feat: humanChatPage에서 ProfileSetting page 분리
