+++ frontend/src/App.tsx
@@ -2,8 +2,7 @@ import { Route, Routes } from 'react-router-dom';
 
 import AIChatPage from './pages/AIChatPage';
 import HomePage from './pages/HomePage';
-import HumanChatPage from './pages/HumanChatPage';
-import ReaderListPage from './pages/ReaderListPage';
+import HumanChatPage, { ChattingPage, SettingPage } from './pages/HumanChatPage';
 import ResultSharePage from './pages/ResultSharePage';
 
 import BackgroundMusic from './components/BackgroundMusic';
@@ -25,11 +24,16 @@ function App() {
         <Route
           path="/chat/human/:roomName"
           element={<HumanChatPage />}
-        />
-        <Route
-          path="/readers"
-          element={<ReaderListPage />}
-        />
+        >
+          <Route
+            path="/chat/human/:roomName"
+            element={<ChattingPage />}
+          />
+          <Route
+            path="/chat/human/:roomName/setting"
+            element={<SettingPage />}
+          />
+        </Route>
         <Route
           path="/result/:id"
           element={<ResultSharePage />}
