+++ frontend/src/pages/HumanChatPage/HumanChatPage.tsx
@@ -1,58 +1,18 @@
-import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
 import Background from '@components/Background';
-import CamContainer from '@components/CamContainer';
 import ChatContainer from '@components/ChatContainer';
 import Header from '@components/Header';
-import ProfileSetting from '@components/ProfileSetting';
 import SideBar from '@components/SideBar';
 
-import useOverlay from '@business/hooks/useOverlay';
 import { useWebRTC } from '@business/hooks/useWebRTC';
+
+export type OutletContext = ReturnType<typeof useWebRTC>;
 
 export default function HumanChatPage() {
   const { roomName } = useParams();
-  const { open } = useOverlay();
-  const [setting, setSetting] = useState(true);
-
-  const {
-    cameraOptions,
-    localVideoRef,
-    remoteVideoRef,
-    toggleAudio,
-    toggleVideo,
-    changeCamera,
-    cameraConnected,
-    changeVideoTrack,
-  } = useWebRTC(roomName as string);
-
-  const changeMyCamera = async (deviceId: string) => {
-    await changeCamera(deviceId);
-    changeVideoTrack();
-  };
-
-  const openProfileSetting = () => {
-    setSetting(true);
-    const camList = cameraOptions.map(({ deviceId, label }) => ({ label, value: deviceId }));
-
-    open(({ close }) => (
-      <ProfileSetting
-        close={() => {
-          close();
-          setSetting(false);
-        }}
-        toggleVideo={toggleVideo}
-        toggleAudio={toggleAudio}
-        cameraConnected={cameraConnected}
-        changeMyCamera={changeMyCamera}
-        camList={camList}
-        videoRef={localVideoRef}
-      />
-    ));
-  };
-
-  useEffect(() => openProfileSetting(), []);
+  const webRTCData = useWebRTC(roomName as string);
 
   return (
     <Background type="dynamic">
@@ -67,15 +27,7 @@ export default function HumanChatPage() {
           </SideBar>,
         ]}
       />
-      {!setting && (
-        <CamContainer
-          localVideoRef={localVideoRef}
-          remoteVideoRef={remoteVideoRef}
-          toggleVideo={toggleVideo}
-          toggleAudio={toggleAudio}
-          cameraConnected={cameraConnected}
-        />
-      )}
+      <Outlet context={webRTCData} />
     </Background>
   );
 }
