+++ frontend/src/pages/HumanChatPage/index.ts
@@ -1,1 +1,4 @@
 export { default } from './HumanChatPage';
+
+export { default as ChattingPage } from './ChattingPage';
+export { default as SettingPage } from './SettingPage';
