--- frontend/src/pages/ReaderListPage/ReaderListPage.tsx
@@ -1,7 +1,0 @@
-interface ReaderListPageProps {}
-
-function ReaderListPage({}: ReaderListPageProps){
-  return <></>;
-};
-
-export default ReaderListPage;
